+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [0.3.0] - IN PROGRESS
 ### Added
 - Add status code and headers in responses
+- Add retry in the reverse proxy forward method
 
 ## [0.1.0] - 2019-01-24
 ### Added